Hotfix attempt: properly display kubectl client version
In the previous version, for some reason, the kubectl client version
is not displayed. This is shown instead:

   source "/pipeline/kubectl-333644c5-def8-4347-8fd3-7ebfecbbb96f/run.sh" < /dev/null
   Running kubectl version:
   Please enter Username: error: EOF

For some reason, it's like kubectl is trying to run some command that
tries to reach the apiserver, which means it's trying to run something
else than `kubectl version --client` (and I'm not sure what that might
be).

So, for now, I'm running the version command in the same shell as the
rest, with the same global parameters, to see what happens.